module functions refactoring
