+++ tests/regtest/1_bitcoin/3_blockchainApplication/output.xml
@@ -33,11 +33,11 @@
   <plugin id="txInjector" path="transaction.so"/>
   <!-- the length of our experiment in seconds  1404101800-->
   <!-- our nodes -->
-  <kill time="50"/>
+  <kill time="10"/>
   <node id="bcdnode0" iphint="1.0.0.1">
     <application arguments="-debug -datadir=data/bcdnode0 -port=18333 -txindex=1 -fallbackfee=0.0002 -rpcuser=a -rpcpassword=1234 -rpcport=11111 -rpcallowip=1.0.0.1/0 -rpcbind=1.0.0.1 -addnode=1.0.0.1:18333 -addnode=1.1.0.1:18333 -algorithm=pow -difficulty=3" plugin="bitcoind" time="0"/>
   </node>
   <node id="client0">
-    <application arguments="1.0.0.1:11111 44 " plugin="client" time="5"/>
+    <application arguments="1.0.0.1:11111 4 " plugin="client" time="5"/>
   </node>
 </shadow>
