+++ tests/regtest/1_bitcoin/7_mainchain/mainchaintest.py
@@ -3,7 +3,7 @@ from subprocess import check_output
 import argparse
 import sys
 sys.path.append("./../../../../")
-from testlibs import test_modules
+from testlibs import test_modules, utils, xml_modules
 
 # Get "bestblockchash" value info using rpc.output_file then check for the same value in bitcoin output log.
 def test_MainChainInfo(shadow_output_file, rpc_output_file):
@@ -49,24 +49,24 @@ def main():
     OPT_REGTEST = args.regtest
     
     if OPT_REGTEST:
-        test_modules.get_xmlfile(path)
+        xml_modules.get_xmlfile(path)
 
     # xml 파일 생성 확인
     target_folder_xml = test_modules.test_xml_existence("output.xml")
 
     # 생성된 xml 파일로 부터 runtime, node_id, plugin 들을 뽑아옴.
-    runtime, node_id_list, plugin_list = test_modules.get_xml_info_new(target_folder_xml)
+    runtime, node_id_list, plugin_list = xml_modules.get_xml_info_new(target_folder_xml)
 
     # 지금 예제는 transaction injector를 사용 안하기에 별도의 xml파일을 만들어줌.
     # 기존 xml에는 transaction.so에 대한 정의가 있어서, 이를 삭제안해주면 shadow가 실행이 안됨.
-    test_modules.set_plugin_file(len(node_id_list), path)
-    test_modules.remove_tx_plugin(target_folder_xml)
+    utils.set_plugin_file(len(node_id_list), path)
+    utils.remove_tx_plugin(target_folder_xml)
     target_folder_xml = target_folder_xml[:len(target_folder_xml)-4] + "2.xml"
     shadow_command = "shadow " + target_folder_xml
 
     # shadow 실행
     print("shadow running ...")
-    test_modules.subprocess_open('shadow output2.xml > output.txt')
+    utils.subprocess_open('shadow output2.xml > output.txt')
     output_file = path + "/output.txt" 
 
     # shadow output 파일 존재 검사.
